vold: tunning the min_rario of fuse
The writing thread will running slowly because of rate-limiting when
proportion of fuse-bdi is near-zero... The proportion of fuse-bdi
will also ramp up slowly because the side effects of performance
counter which can be upto number_of_cpus * counter_batch away from
the real counter value.

To prevent this,tunning min_ratio to give the bdi some breathing
room for ramping up.

Change-Id: I71b7f0e34c05cc75427a576cc7771633ef726d15
Signed-off-by: Chi Wu <wuchi.zero@gmail.com>